Fixed 2 warning messages (werror build)   * Fix a cast error (only shows at x86_64)   * Missing include, added gfpm-interface.h to quickpane.c to fix a missing     declaration

+++ src/gfpm-interface.c
@@ -923,7 +923,7 @@ gfpm_trans_prepare (PM_LIST *list)
 		str = g_strdup_printf (_("Failed to prepare transaction (%s)\n"), pacman_strerror (pm_errno));
 		gfpm_error (_("Error"), str);
 		g_free (str);
-		int t = pacman_trans_getinfo (PM_TRANS_TYPE);
+		int t = (long)pacman_trans_getinfo (PM_TRANS_TYPE);
 		switch ((long)pm_errno)
 		{
 			case PM_ERR_UNSATISFIED_DEPS:
